+++ Constructor/Constructor/View/Items/ItemDetailsView.swift
@@ -61,7 +61,7 @@ extension ItemDetailsView {
     
     var buttons: some View {
         Button {
-            superVM.updateItem(item: item, name: name, price: Float(price) ?? 0)
+            superVM.updateItem(item: item, name: name, price: Double(price) ?? 0)
             superVM.saveItem()
             presentationMode.wrappedValue.dismiss()
         } label: {
